+++ branches/dev_src/src/interaction.h
@@ -14,8 +14,8 @@
 void COLL_compute_phi_gradients(void);
 void COLL_compute_phi_fluid(void);
 
-Colloid * COLL_add_colloid(int, Float, Float, FVector, FVector, FVector);
-Colloid * COLL_add_colloid_no_halo(int, Float, Float, FVector, FVector,
+Colloid * COLL_add_colloid(int, double, double, FVector, FVector, FVector);
+Colloid * COLL_add_colloid_no_halo(int, double, double, FVector, FVector,
 				   FVector);
 void      COLL_zero_forces(void);
 void      COLL_init(void);
